fix(fe): editor isn't reset when creating new note
With this fix, the React app does not possess any state information about the note content blocks anymore. Creating a new note or loading one from the database will immediately trigger an Editor update. The other way round, only note save requests by the user will call Editor.save(). All in all, this means that the single source of truth of the editor content is now the editor/the DOM itself and not React. This should finally fix some bugs regarding the editor not responding correctly to user input.