fix(backend): more precise note feature evaluation
If a note contains text is now not only dependent on if it has a paragraph block, but also if this block is not empty, or if there is a non-empty heading or list block.
The file features are only true if there is a loaded file inside the block.
However, containsText is still true if there are whitespace HTML entities in the text.