chore: adjust tsconfigs to prevent type collisions
Previously, we had one tsconfig.json for the whole repo. This led to type collisions. E. g. ReadableStream is available in the dom as well as in NodeJS types. With the new tsconfigs for specific domains (backend, browser extension, frontend), we inherit from a base config and add only the types required for that domain.