feat: remove refresh note list button
The note list refresh button is currently not useful since we don't have a server state anymore that can change independent of the frontend. Well, we now have a disk state that can change independently, but we currently don't re-parse notes from disk when clicking on this button. And this functionality can technically also be achieved by using the browser's refresh button.