feat: don't destroy graph metadata if disk load throws
Previously, if graph loading from disk throwed, we assumed there is nothing to work with and created a new .graph.json file (even when we throwed for unrelated reasons, e.g. the web worker was unavailable). With this improvement, we just throw and don't catch. But also, if a metadata file is missing now, we just create it at the spot.