feat: don't update graph metadata on note update
With this change, the `updatedAt` timestamp of `.graph.json` won't be updated on every note update anymore, but only on updates in the `.graph.json` itself (e.g. files, pins). If the user wants to retrieve the latest graph update globally via getStats(), it is now calculated based on the latest updated note. Also, the `createdAt` timestamp that is returned by getStats() is now calculated based on the oldest item in the graph.