feat(BREAKING): save SG files for AG files
This breaking change adjusts the way arbitrary graph files (like `image.png`) are saved. Before, we had a graph metadata JSON file that contained all data that could not be stored elsewhere (pins, file metadata, ...). Now, there is a Subtext Graph specification that NENO adheres to. Instead of saving it all to one JSON file, metadata for an arbitrary graph file is stored in a separate Subtext graph file (see spec for more details). This means that from now on, every slug that is used in the graph maps to one Subtext graph file.